Split settings navigation helpers
